# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LOOP API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/audit": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Audit",
        "operationId": "audit_api_v1_admin_audit_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEventRead"
                  },
                  "title": "Response Audit Api V1 Admin Audit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/schedules": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Schedules",
        "operationId": "schedules_api_v1_admin_schedules_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScheduledTaskRead"
                  },
                  "type": "array",
                  "title": "Response Schedules Api V1 Admin Schedules Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Schedule",
        "operationId": "create_schedule_api_v1_admin_schedules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduledTaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledTaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/schedules/{schedule_id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Patch Schedule",
        "operationId": "patch_schedule_api_v1_admin_schedules__schedule_id__patch",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Schedule Id"
            }
          },
          {
            "name": "enabled",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Enabled"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledTaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/schedules/{schedule_id}/run-now": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Run Schedule Now",
        "operationId": "run_schedule_now_api_v1_admin_schedules__schedule_id__run_now_post",
        "parameters": [
          {
            "name": "schedule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Schedule Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRunRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/sales-strategies/{strategy_id}/progress/reconcile": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Reconcile",
        "operationId": "reconcile_api_v1_admin_sales_strategies__strategy_id__progress_reconcile_post",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Reconcile Api V1 Admin Sales Strategies  Strategy Id  Progress Reconcile Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/dead-letters/{dead_letter_id}/replay": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Replay",
        "operationId": "replay_api_v1_admin_dead_letters__dead_letter_id__replay_post",
        "parameters": [
          {
            "name": "dead_letter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dead Letter Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/JobRunRead"
                    },
                    {
                      "$ref": "#/components/schemas/IntegrationEventRead"
                    }
                  ],
                  "title": "Response Replay Api V1 Admin Dead Letters  Dead Letter Id  Replay Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/dead-letters/{dead_letter_id}/discard": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Discard",
        "operationId": "discard_api_v1_admin_dead_letters__dead_letter_id__discard_post",
        "parameters": [
          {
            "name": "dead_letter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dead Letter Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/sales-strategies/{strategy_id}/memory": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Remember",
        "operationId": "remember_api_v1_admin_sales_strategies__strategy_id__memory_post",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrainMemoryCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrainMemoryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Recall",
        "operationId": "recall_api_v1_admin_sales_strategies__strategy_id__memory_get",
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Id"
            }
          },
          {
            "name": "agent_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Type"
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BrainMemoryRead"
                  },
                  "title": "Response Recall Api V1 Admin Sales Strategies  Strategy Id  Memory Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AuditEventRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "before": {
            "additionalProperties": true,
            "type": "object",
            "title": "Before"
          },
          "after": {
            "additionalProperties": true,
            "type": "object",
            "title": "After"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "actor",
          "action",
          "entity_type",
          "entity_id",
          "before",
          "after",
          "reason",
          "request_id",
          "created_at"
        ],
        "title": "AuditEventRead"
      },
...
      "BrainMemoryCreate": {
        "properties": {
          "agent_type": {
            "type": "string",
            "title": "Agent Type"
          },
          "category": {
            "type": "string",
            "enum": [
              "actions",
              "failures",
              "decisions",
              "insights"
            ],
            "title": "Category"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content"
          },
          "evidence_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Urls"
          }
        },
        "type": "object",
        "required": [
          "agent_type",
          "category",
          "content"
        ],
        "title": "BrainMemoryCreate"
      },
      "BrainMemoryRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sales_strategy_id": {
            "type": "string",
            "title": "Sales Strategy Id"
          },
          "agent_type": {
            "type": "string",
            "title": "Agent Type"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "evidence_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Urls"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sales_strategy_id",
          "agent_type",
          "category",
          "content",
          "evidence_urls",
          "created_at"
        ],
        "title": "BrainMemoryRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IntegrationEventRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "event_type",
          "version",
          "payload",
          "correlation_id",
          "published_at",
          "attempts",
          "created_at"
        ],
        "title": "IntegrationEventRead"
      },
      "JobRunRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "task_key": {
            "type": "string",
            "title": "Task Key"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_key",
          "status",
          "attempts",
          "payload",
          "error",
          "created_at"
        ],
        "title": "JobRunRead"
      },
...
      "ScheduledTaskCreate": {
        "properties": {
          "key": {
            "type": "string",
            "minLength": 1,
            "title": "Key"
          },
          "interval_seconds": {
            "type": "integer",
            "maximum": 86400.0,
            "minimum": 60.0,
            "title": "Interval Seconds"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "overlap_policy": {
            "type": "string",
            "enum": [
              "skip",
              "queue"
            ],
            "title": "Overlap Policy",
            "default": "skip"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "key",
          "interval_seconds"
        ],
        "title": "ScheduledTaskCreate"
      },
      "ScheduledTaskRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "interval_seconds": {
            "type": "integer",
            "title": "Interval Seconds"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "overlap_policy": {
            "type": "string",
            "title": "Overlap Policy"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "next_run_at": {
            "type": "string",
            "format": "date-time",
            "title": "Next Run At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "key",
          "interval_seconds",
          "enabled",
          "overlap_policy",
          "payload",
          "next_run_at"
        ],
        "title": "ScheduledTaskRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}